musicapp-django/requirements/base.txt
certifi==2022.12.7
chardet==5.1.0
Click==8.1.3
coreapi==2.3.3
coreapi-cli==1.0.9
coreschema==0.0.4
Django==4.1.5
django-environ==0.9.0
djangorestframework==3.14.0
idna==3.4
itypes==1.2.0
Jinja2==3.1.2
Markdown==3.4.1
MarkupSafe==2.1.1
Pillow==9.4.0
Pygments==2.14.0
pytz==2019.2
requests==2.28.1
six==1.16.0
sqlparse==0.4.3
uritemplate==4.1.1
urllib3==1.26.13

asgiref==3.6.0
certifi==2022.12.7
chardet==5.1.0
charset-normalizer==2.1.1
click==8.1.3
coreapi==2.3.3
coreapi-cli==1.0.9
coreschema==0.0.4
Django==4.1.5
django-environ==0.9.0
djangorestframework==3.14.0
idna==3.4
itypes==1.2.0
Jinja2==3.1.2
Markdown==3.4.1
MarkupSafe==2.1.1
Pillow==9.4.0
Pygments==2.14.0
pytz==2019.2
requests==2.28.1
six==1.16.0
sqlparse==0.4.3
uritemplate==4.1.1
urllib3==1.26.13
